+++ lib/globals.dart
@@ -17,8 +17,8 @@ final GlobalKey<ScaffoldMessengerState> scaffoldMessengerKey = GlobalKey<Scaffol
 final GlobalKey<NavigatorState> navigatorKey = GlobalKey<NavigatorState>();
 
 const String prgName = 'easysend';
-const String progVersion = '0.1.260807';
-const int buildNumber = 23;
+const String progVersion = '0.1.260808';
+const int buildNumber = 28;
 const String progAuthor = 'Eugen';
 
 const String langFile = 'assets/locales.json';